Main (#2)
* fix: replace base64 decoding with jq built in base64 decoding for compatibility (#1)

thanks @jeff-auth0

* fix: make execution of .env file consistent across files (#2)

* chore: Add MIT license info

* chore: rebase and fixes with main branch

* refactor: remove .env logic as new logic will be added soon

* refactor: move shabang to top level

* Add access token scope check - Part 1

* Support fetching connection cert

Also removes unsupported /_release endpoint

* Add access token scope check - Part 2

* refactor: remove -u flag

refactor: format error lines

* Add Actions support

* use bash from env

use bash from env
use command vs which

DIR clean up

management api access_token script for cli export

body declare cleaner

* Support changes from #11 in Actions

Also adds verbose mode to all Actions scripts

* Add attack protection support (#15)

* fix indentation issue from vscode plugin (#16)

* move sign-up to users (#17)

* support for guardian push notification in mf cli (#18)

* Feat/guardian sns (#19)

* first part of guardian push notification cli - scanner & enrollment

* improve some .env imports

+++ account-linking/unlink.sh
@@ -1,6 +1,15 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
-set -euo pipefail
+##########################################################################################
+# Author: Auth0
+# Date: 2022-06-12
+# License: MIT (https://github.com/auth0/auth0-bash/blob/main/LICENSE)
+##########################################################################################
+
+set -eo pipefail
+
+command -v curl >/dev/null || { echo >&2 "error: curl not found";  exit 3; }
+command -v jq >/dev/null || {  echo >&2 "error: jq not found";  exit 3; }
 
 function usage() {
     cat <<END >&2
@@ -8,7 +17,7 @@ USAGE: $0 [-e env] [-a access_token] [-p primary] [-s secondary] [-v|-h]
         -e file        # .env file location (default cwd)
         -a token       # Access Token
         -i user_id     # primary user_id
-        -p provider    # second user's provider 
+        -p provider    # second user's provider
         -s user_id     # secondary user_id
         -h|?           # usage
         -v             # verbose
@@ -24,26 +33,34 @@ declare userId=''
 declare provider=''
 declare opt_verbose=0
 
-while getopts "e:a:i:p:s:hv?" opt
-do
+while getopts "e:a:i:p:s:hv?" opt; do
     case ${opt} in
-        e) source ${OPTARG};;
-        a) access_token=${OPTARG};;
-        i) userId=${OPTARG};;
-        p) provider=${OPTARG};;
-        s) secondary_userId=${OPTARG};;
-        v) opt_verbose=1;; #set -x;;
-        h|?) usage 0;;
-        *) usage 1;;
+    e) source ${OPTARG} ;;
+    a) access_token=${OPTARG} ;;
+    i) userId=${OPTARG} ;;
+    p) provider=${OPTARG} ;;
+    s) secondary_userId=${OPTARG} ;;
+    v) opt_verbose=1 ;; #set -x;;
+    h | ?) usage 0 ;;
+    *) usage 1 ;;
     esac
 done
 
-[[ -z "${access_token}" ]] && { echo >&2 "ERROR: access_token undefined"; usage 1; }
-[[ -z "${userId}" ]] && { echo >&2 "ERROR: userId undefined"; usage 1; }
-[[ -z "${secondary_userId}" ]] && { echo >&2 "ERROR: secondary_userId undefined"; usage 1; }
-[[ -z "${provider}" ]] && { echo >&2 "ERROR: provider undefined"; usage 1; }
+[[ -z "${access_token}" ]] && { echo >&2 "ERROR: access_token undefined";  usage 1; }
 
-declare -r AUTH0_DOMAIN_URL=$(echo ${access_token} | awk -F. '{print $2}' | base64 -di 2>/dev/null | jq -r '.iss')
 
-curl -X DELETE  -H "Authorization: Bearer $access_token" \
+declare -r AVAILABLE_SCOPES=$(jq -Rr 'split(".") | .[1] | @base64d | fromjson | .scope' <<< "${access_token}")
+declare -r EXPECTED_SCOPE="update:users"
+[[ " $AVAILABLE_SCOPES " == *" $EXPECTED_SCOPE "* ]] || { echo >&2 "ERROR: Insufficient scope in Access Token. Expected: '$EXPECTED_SCOPE', Available: '$AVAILABLE_SCOPES'"; exit 1; }
+
+[[ -z "${userId}" ]] && { echo >&2 "ERROR: userId undefined";  usage 1; }
+
+[[ -z "${secondary_userId}" ]] && { echo >&2 "ERROR: secondary_userId undefined";  usage 1; }
+
+[[ -z "${provider}" ]] && { echo >&2 "ERROR: provider undefined";  usage 1; }
+
+
+declare -r AUTH0_DOMAIN_URL=$(jq -Rr 'split(".") | .[1] | @base64d | fromjson | .iss' <<<"${access_token}")
+
+curl -X DELETE -H "Authorization: Bearer $access_token" \
     --url "${AUTH0_DOMAIN_URL}api/v2/users/${userId}/identities/${provider}/${secondary_userId}"
